Update check improvements
